+++ NEWS.rst
@@ -5,7 +5,7 @@ News
 Version 0.4.1 - 2014-..-..
 
   * Beta status
-  * added convenience methods place() and grid() to the Insert entity
+  * added convenience methods place(), grid(), get_attrib_text() and has_attrib() to the Insert entity
 
 Version 0.4.0 - 2014-03-16